requirements.txt
bcrypt==3.1.4
cffi==1.11.4
Django==1.11
ez-setup==0.9
pycparser==2.18
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
